Completed conveyer belt physics, although this reveals an issue where slamming into a wall at high speed causes you to 'land' in it.
Fixed an issue where it was almost impossible to jump through 2 thru
blocks flanked by solid blocks without pixel-perfect precision. Added a
'special case' that 'snaps' bonzo in place to do the jump if he is close
enough.

Changed bonzos speed from 2 to 3. This may be a bit fast. The ideal
speed would be 2.5 but it seams some integer conversion is preventing
incremental speeds of this nature (he would move faster to the left than
to the right due to truncation). Must find source of truncation so his
speed can be tuned.