UI added plus new resources!
Added main banner, number bitmaps for score and bonuses, and finally the
explosions whilst I was at it. Considering implementing health as a
dynamically drawn and rendered rectangle vs a sprite.

Only the UI draws. Need to actually implement concepts that are
reflected on UI before it becomes more useful.