Added more patterns to make issue of no scroll bars apparent in debug world even at refined size.
Found a long-standing bug where there could only be a max of 10 (0-9)
types of sprites, backgrounds, or patterns because ArrayList.add (and
LinkedList.add) can't handle adding to a specific index if the size
isn't big enough. I was under the wrong assumption that they did that.
Converted to basic arrays of size 256 for now, but eventually wish to
replace with a class that can handle what I need without a hard-coded
upper bound.