Added music to worlds; a music.ogg file in the .zip will allow background music. No such file being found is still okay. Fixes #4
Updated debug world to include music for testing

Fixed an issue where, when a level or screen doesn't exist to the side
and the game engine needs to 'pacman' bonzo, he would both die if he
jumped through the 'pacman' effect, or, and more importantly, the game
would freeze if bonzo tried to jump after walking through one of these
effects. This is mainly to prevent crashes when entering worlds not
designed with levels in certain directions.

Updated .jar