Fixed the off by one errors, but sprites are not picking up on two-sets horizontal (though they are on slow animation).
Fixed id inversion. Thanks to a little mistake I made when I first
started this project, translating the original level ids to ids for the
port requires inverting the value (+100 for going up in port, but -100
in original). Current method removes the entire 900-series row from the
original levels; will need to fix