Fixed Okay Button.
Applause would not play because the resources would be disposed and
closed before the playback would finish. Added code that allowed the
clip to be 'held on' until being finished and have a late release, and
adjusted timing to best work with the situation.

Fixed high score name entry dialog to be a more reasonable size.

Still some rough edges but otherwise represents the concept of high
scores.