Added location information in Sprite selector when there is more than one (both for editing offscreen sprites and editing with a click when multiple sprites are in the region).
Changed logic for replacing sprites. If a sprite that is being edited is
at index n in the internal list, it is removed and the new sprite is
added at n using the insertion list iterator. This acts as a replace and
maintains order with other sprites. If more than 1 sprite have the same
icon and location, and even when they don't, when the sprite chooser
dialog pops up, having them always displayed in a consistent order is
important.