Refactored again to use a save state object.
This preserves the jump state, and I also noticed that conveyer state
would need to be preserved as well. The last thing is Bonzo's current
animation. Logically, the game runs as it should, but bonzo looks like
he starts a jump in midair if he respawns in air (although all the
proper jump physics do now kick in)