Disabled Display List caching for main font.
Guessed (turns out hopefully correctly) that the font drawing issue was
similar to the texture issue when the texture cache is not deleted upon
loading a new world. Font drawing now works in multiple game sessions
without having to restart the application.

As of now, full function (if looking a bit different in some places) is
restored, just using Slick2D. This can be merged into master.