Removed Maven stuff/adding Gradle stuff
This also removes the build distribution .exes.

The intent is to eventually let Gradle build a far jar automatically,
set up a release with native libraries in the right place, and the
system path that must be set for the libraries to be loaded passed to
both the game and level editor executables.

Unsure if the last part is going to even be realistically automatable in
the time I have, but at the very least the fat jar needs to be created
(it can be) and sent into a folder where native libraries exist (todo).