Split Sprite into Monster and Sprite classes.
Sprite is now abstract with the intention of eventually allowing Goodies
and Hazards to inherit. This is primarily to re-use the collision
detection.

The current split is rough; did a naive copy-paste in many places to use
the new Monster class, but some algorithms may be able to standardise on
just the Sprite class (such as drawing and collision detection).